==> scripts/00-data-analysis/B-diveboard-data/05b_mpa_lit_comparison.Rmd <==
---
output:
  html_document: default

title: 'Dive Database: Compare number of extrapolated dives to literature estimates (MPA Level)'
author: "Kat Millage"
date: "`r format(Sys.time(), '%B %d, %Y')`"
fontsize: 11pt
---

```{r setup, include=FALSE}
# Packages
library(readxl)

# Source common R file
source(here::here("common.R"))
```

# Introduction

This script compares our extrapolated estimates of the number of dives made in different locations to estimates from literature. This is the MPA-level comparison.  

```{r}
# Load the dive data matched to site.
dive_dat <- read_csv(file.path(emlab_project_dir, "data", "02-processed", "data-prep", "diveboard_database_extrapolated_2010_2020_final.csv"))

# Load the dataset from literature
lit_dat_raw <- readxl::read_xlsx(file.path(emlab_project_dir, "data", "01-raw", "literature", "diving_dat_from_literature.xlsx")) %>%
  janitor::clean_names()
```

# Plot Settings & Setup

```{r}
# Color pallette so it's consistent across both National and MPA analysis 
pal_names <- unique(lit_dat_raw$region)
color_pal <- hue_pal()(length(pal_names))
names(color_pal) <- pal_names
```

# Get Protected Areas of Interest Totals from Literature

```{r}
lit_dat_mpas <- lit_dat_raw %>%
    dplyr::filter(scale == "MPA" & !is.na(wdpa_id) & sharks_only == "N")

# Get unique MPAs
mpa_ids <- unique(lit_dat_mpas$wdpa_id)
```

# Get Protected Area Totals from Extrapolated Database

Even though we already overlaid the protected area boundaries on our database, we only kept the area with the highest protection status. Unfortunately, it's not uncommon for the most highly protected area to only be a subset of the total protected area, and the estimates from literature generally tend to correspond to the total protected area. Thus, they doesn't match up perfectly if we just summarize by the WDPA ID's already assigned to our dive database. So we'll overlay the protected area boundaries again, but only for the IDs we have from literature. 

## Turn site coordinates into spatial points

```{r}
# Extract the spot IDs and lat/lon
site_coords <- dive_dat %>%
  distinct(spot_id, lat, lon)

# Make coordinates into spatial points
site_coords_sf <- site_coords %>%
  st_as_sf(coords = c("lon", "lat"),
           remove = T)
```

## Load MPA boundaries, overlay dive sites, and calculate dives

```{r}
mpa_dat_extrapolated <- tibble(0)

for(i in 1:length(mpa_ids)){
  
### Load MPA .gdb file from WDPA
mpa_code <- mpa_ids[i]

mpa_dat_path <- file.path(emlab_project_dir, "data", "01-raw", "wdpa", paste0("WDPA_WDOECM_Jun2022_Public_", mpa_code), paste0("WDPA_WDOECM_Jun2022_Public_", mpa_code, ".gdb"))
mpa_layers <- ogrListLayers(mpa_dat_path)

### Read in MPA boundaries
mpa_sf <- st_read(mpa_dat_path, mpa_layers[1]) %>%
  janitor::clean_names()

### Convert our coords
st_crs(site_coords_sf) <- st_crs(mpa_sf)

### Find sites within the MPA boundaries
sites_in_mpa <- st_join(site_coords_sf, mpa_sf, join = st_within) %>%
  dplyr::filter(!is.na(wdpaid)) %>%
  st_drop_geometry() %>%
  distinct(spot_id, wdpaid, wdpa_pid, name, iucn_cat, marine, no_take)

### Match back to dive dataset
matched_sites <- dive_dat %>%
  right_join(sites_in_mpa, by = "spot_id") %>%
  group_by(wdpaid, wdpa_pid, name, iucn_cat.y, marine, no_take) %>%
  summarize(n_dives = n_distinct(dive_id),
            n_dives_extrap = sum(n_dives_extrap, na.rm = T),
            n_dives_extrap_min = sum(n_dives_extrap_min, na.rm = T),
            n_dives_extrap_max = sum(n_dives_extrap_max, na.rm = T)) %>%
  ungroup()

mpa_dat_extrapolated <- mpa_dat_extrapolated %>%
  bind_rows(matched_sites)

}

# Extract what we need
mpa_dat_extrapolated_out <- mpa_dat_extrapolated %>%
  group_by(wdpaid, name) %>%
  summarize(n_dives_extrap = sum(n_dives_extrap, na.rm = T),
            n_dives_extrap_min = sum(n_dives_extrap_min, na.rm = T),
            n_dives_extrap_max = sum(n_dives_extrap_max, na.rm = T)) %>% # summing parts of the great barrier reef
  ungroup() %>%
  rename(wdpa_id = wdpaid) %>%
  dplyr::filter(!is.na(wdpa_id))
```

# Get Protected Area Totals from Literature

```{r}
# Let's extract the entries we have for number of dives - we don't need to do anything to these
dives_lit <- lit_dat_mpas %>%
  dplyr::filter(variable == "Dives") %>%
  dplyr::select(id, region, country_territory, wdpa_id, value, min, max)

# Now let's extract the entries we have for number of divers - we're going to make a couple of assumptions here. 
# Let's assume the average diver makes anywhere between 2 and 4 dives per year to give us a range. 
# We're only going to keep this extrapolation for sources where we don't have a direct estimate of the number of dives
divers_lit <- lit_dat_mpas %>%
  dplyr::filter(variable == "Divers") %>%
  mutate(dives_min = case_when(!is.na(value) ~ value*2,
                               T ~ min*2),
         dives_max = case_when(!is.na(value) ~ value*4,
                               T ~ max*4)) %>%
  dplyr::filter(!(id %in% dives_lit$id)) %>%
  dplyr::select(id, region, country_territory, wdpa_id, min = dives_min, max = dives_max)

# Now we join them back together and select which to keep
dives_lit_out <- dives_lit %>%
  bind_rows(divers_lit)

# Now we summarize by protected area 
dives_lit_out <- dives_lit_out %>%
  group_by(region, country_territory, wdpa_id) %>%
  summarize(value_min = min(value, na.rm = T),
            value_max = max(value, na.rm = T),
            range_min = min(min, na.rm = T),
            range_max = max(max, na.rm = T)) %>%
  ungroup() %>%
  rowwise() %>%
  mutate(min = min(c(value_min[value_min != Inf], range_min[range_min != Inf])),
         max = max(c(value_max[value_max != -Inf], range_max[range_max != -Inf])),
         median = median(c(min, max))) %>%
  dplyr::select(region, country_territory, wdpa_id, min, max, median) %>%
  mutate(country_iso3 = countrycode(country_territory, "country.name", "iso3c")) %>%
  ungroup()
```

# Merge

```{r}
# Join extrapolation
dives_matched <- dives_lit_out %>%
  left_join(mpa_dat_extrapolated_out, by = "wdpa_id") %>%
  rename(dives_extrap = n_dives_extrap)

# Deal with Bonaire and Saba
dives_matched$country_territory[dives_matched$country_territory %in% c("Bonaire", "Netherlands Antilles")] <- "Bonaire, Sint Eustatius and Saba"
dives_matched$country_iso3[dives_matched$country_territory %in% c("Bonaire, Sint Eustatius and Saba")] <- "BES"

write_csv(dives_matched, file.path(emlab_project_dir, "data", "diagnostics", "mpa_lit_comparison.csv"))
```

## Plot

```{r}
plot_dat <- dives_matched

max_y <- max(c(plot_dat$max, plot_dat$dives_extrap))*1.05

# Plot
dives_point_plot <- plot_dat %>%
  ggplot(aes(x = dives_extrap, y = median, color = region))+
  geom_point(size = 2)+
  geom_crossbar(aes(x = dives_extrap, ymin = min, ymax = max), width = 0.5)+
  geom_crossbar(aes(y = median, xmin = n_dives_extrap_min, xmax = n_dives_extrap_max), width = 0.5, color = "black", alpha =0.5)+
  geom_abline(slope = 1, intercept = 0, linetype = "dashed", color = "lightgrey")+
  scale_y_continuous(limits = c(0,max_y))+
  scale_x_continuous(limits = c(0,max_y))+
  labs(x = "Dives per Year (#, extrapolated)", y = "Dives per Year (#, from literature)", color = "")+
  scale_color_manual(values = color_pal[names(color_pal) %in% plot_dat$region],
                     guide = guide_legend(title.position = "top",
                                          title.hjust = 0.5,
                                          nrow = 2,
                                          byrow = T))+
  theme_basic()

# Output
save_plots(dives_point_plot, 
           file_name = file.path(emlab_project_dir, "figures", "lit-comparison", "dives_extrap_lit_comparison_point_plot_mpa.png"),
           width_in = 7.5,
           height_in = 5.5)

dives_point_plot
```

```{r}
# Let's also make a bar plot version
dives_bar_plot <- plot_dat %>%
  ggplot()+
  aes(x = fct_rev(as.factor(name)))+
  geom_segment(aes(xend = name, y = min, yend = max, color = region), size = 4, alpha = 0.75)+
  geom_point(aes(y = median, color = region), size = 5)+
  geom_segment(aes(xend = name, y = n_dives_extrap_min, yend = n_dives_extrap_max), color = "black", size = 1, alpha = 0.5)+
  geom_point(aes(y = dives_extrap), size = 2)+
  labs(x = "", y = "Dives per Year (#)", color = "")+
  coord_flip()+
  scale_color_manual(values = color_pal[names(color_pal) %in% plot_dat$region],
                     guide = guide_legend(title.position = "top",
                                          title.hjust = 0.5))+
  theme_basic()

# Output
save_plots(dives_bar_plot, 
           file_name = file.path(emlab_project_dir, "figures", "lit-comparison", "dives_extrap_lit_comparison_bar_plot_mpa.png"),
           width_in = 7.5,
           height_in = 5.5)

dives_bar_plot
```

```{r}
dives_plot_combined <- cowplot::plot_grid(dives_bar_plot + theme(legend.position = "none"),
                                          dives_point_plot + theme(legend.position = "none"),
                                          nrow = 1,
                                          align = "h",
                                          rel_widths = c(1.4,0.9))

dives_plot_combined_w_legend <- cowplot::plot_grid(dives_plot_combined,
                                                   cowplot::get_legend(dives_bar_plot),
                                                   nrow = 2,
                                                   align = "v",
                                                   rel_heights = c(1, 0.3))

# Output
save_plots(dives_plot_combined_w_legend, 
           file_name = file.path(emlab_project_dir, "figures", "lit-comparison", "dives_extrap_lit_comparison_mpa_combo_plot.png"),
           width_in = 14,
           height_in = 6.5)

dives_plot_combined_w_legend
```
